+++ sys/netns/ns.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)ns.c	8.2 (Berkeley) 11/15/93
+ *	@(#)ns.c	8.3 (Berkeley) 01/09/95
  */
 
 #include <sys/param.h>
@@ -33,7 +33,7 @@ extern struct sockaddr_ns ns_netmask, ns_hostmask;
 /* ARGSUSED */
 ns_control(so, cmd, data, ifp)
 	struct socket *so;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	register struct ifnet *ifp;
 {
